backend/init/requirements.txt
weaviate-client==4.14.1